+++ livelossplot/outputs/__init__.py
@@ -12,7 +12,6 @@ from .extrema_printer import ExtremaPrinter
 # hack-ish, but works (and I am not aware of a more proper way to do so)
 
 from .bokeh_plot import BokehPlot
-from .neptune_logger import NeptuneLogger
 from .tensorboard_logger import TensorboardLogger
 from .tensorboard_tf_logger import TensorboardTFLogger